criação do projeto web api e alteração do projeto repositorio para uma arquitetura mais solida

+++ NUINFDOMINIO/Pessoa.cs
@@ -6,7 +6,7 @@ namespace NUINFDOMINIO
 {
   public class Pessoa
     {
-        public long id { get; set; }
+        public int id { get; set; }
         public string nome { get; set; }
         public string cpf { get; set; }
         public DateTime dataNascimento { get; set; }
